+++ test/support/factory.ex
@@ -31,7 +31,6 @@ defmodule Pleroma.Factory do
       nickname: sequence(:nickname, &"nick#{&1}"),
       password_hash: Comeonin.Pbkdf2.hashpwsalt("test"),
       bio: sequence(:bio, &"Tester Number #{&1}"),
-      info: %{},
       last_digest_emailed_at: NaiveDateTime.utc_now()
     }